third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_reparenting-expected.txt

This is a testharness.js-based test.
PASS http://127.0.0.1:8000/w3c/webperf/resources/generate_resource.php?type=image&id=move_to_child is expected to have initiatorType img 
Harness: the test ran to completion.
